Modified FFT to avoid window function
Instead of doing a window function its better to extend the  FFT to
twice the length with the complex conjugate. Then after the FFT trow
away the imag part. As the FFT is limited to 128 point I only used the
first 64. Better would be to exten to128 and then af extending with the
complex conjugate do a 256 point FFT